+++ src/main/java/com/wnsud9771/reoisitory/pipeline/GraphPipelinesConnectRepository.java
@@ -1,0 +1,9 @@
+package com.wnsud9771.reoisitory.pipeline;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.wnsud9771.entity.dashboard.GraphPipelinesConnect;
+
+public interface GraphPipelinesConnectRepository  extends JpaRepository<GraphPipelinesConnect, Long>{
+
+}
